move getBookFromStack from BookRightClickHandler to ItemStackUtil
